Remove shell process replacement
Bash on Travis CI insists on reporting as interactive when it's not, so
the detection for an interactive shell process doesn't work quite right.
Instead, check for a new or updated bash install and inform the user
that they need to exit the current shell process to use the new version.